ACTINS-2088 Courses integration control flow for existing file or uploaded file.  Changed arguments in rake task call

+++ app/jobs/lionpath_integrate_job.rb
@@ -1,11 +1,17 @@
 class LionpathIntegrateJob < ApplicationJob
 
-  def perform(params, log_path)
+  # When using rake task, file_exist must = true
+  # This will use existing file 'app/parsing_files/courses_taught.csv' instead of uploaded file
+  def perform(params, log_path, file_exist=false)
     error_logger = Logger.new("public/#{log_path}")
     error_logger.info "Courses Taught Integration to #{params[:target]} initiated at: #{DateTime.now}"
-    f_name = params[:courses_file].original_filename
-    f_path = File.join('app', 'parsing_files', f_name)
-    File.open(f_path, "wb") { |f| f.write(params[:courses_file].read) }
+    if file_exist.false?
+      f_name = params[:courses_file].original_filename
+      f_path = File.join('app', 'parsing_files', f_name)
+      File.open(f_path, "wb") { |f| f.write(params[:courses_file].read) }
+    else
+      f_path = File.join('app', 'parsing_files', 'courses_taught.csv')
+    end
     my_lionpath_populate = LionPathPopulateDB.new(LionPathParser.new(filepath = f_path))
     my_lionpath_populate.format_and_filter
     my_lionpath_populate.populate
